requirements.txt
python-telegram-bot>=21.0
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httpx>=0.25.0
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2023.3
